pulls the mechanism __repr__()/__str__() -> info() -> summary_dict() up to DimcatObject
